move clickhouse processor base with all subclasses
