+++ snuba/datasets/storage.py
@@ -1,7 +1,6 @@
 from abc import ABC, abstractmethod
 from typing import Any, NamedTuple, Optional, Sequence
 
-from snuba.clickhouse.processors import ClickhouseQueryProcessor
 from snuba.clickhouse.translators.snuba.mapping import TranslationMappers
 from snuba.clusters.cluster import (
     ClickhouseCluster,
@@ -16,6 +15,7 @@ from snuba.datasets.storages.storage_key import StorageKey
 from snuba.datasets.table_storage import KafkaStreamLoader, TableWriter
 from snuba.query.expressions import Expression
 from snuba.query.logical import Query
+from snuba.query.processors.physical.processors import ClickhouseQueryProcessor
 from snuba.query.query_settings import QuerySettings
 from snuba.replacers.replacer_processor import ReplacerProcessor
