+++ snuba/pipeline/composite.py
@@ -4,10 +4,6 @@ from typing import Mapping, NamedTuple, Optional, Sequence, Tuple, Union
 
 import sentry_sdk
 
-from snuba.clickhouse.processors import (
-    ClickhouseQueryProcessor,
-    CompositeQueryProcessor,
-)
 from snuba.clickhouse.query import Query as ClickhouseQuery
 from snuba.clusters.cluster import ClickhouseCluster, get_cluster
 from snuba.clusters.storage_sets import StorageSetKey, is_valid_storage_set_combination
@@ -30,6 +26,10 @@ from snuba.query.joins.equivalence_adder import add_equivalent_conditions
 from snuba.query.joins.semi_joins import SemiJoinOptimizer
 from snuba.query.joins.subquery_generator import generate_subqueries
 from snuba.query.logical import Query as LogicalQuery
+from snuba.query.processors.physical.processors import (
+    ClickhouseQueryProcessor,
+    CompositeQueryProcessor,
+)
 from snuba.query.query_settings import QuerySettings
 from snuba.web import QueryResult
