+++ snuba/query/processors/physical/abstract_array_join_optimizer.py
@@ -1,7 +1,6 @@
 from itertools import combinations
 from typing import Callable, Dict, List, Optional, Sequence, Set, Tuple, TypeVar, Union
 
-from snuba.clickhouse.processors import ClickhouseQueryProcessor
 from snuba.clickhouse.query import Query
 from snuba.query.conditions import (
     BooleanFunctions,
@@ -23,6 +22,7 @@ from snuba.query.matchers import (
     Pattern,
     String,
 )
+from snuba.query.processors.physical.processors import ClickhouseQueryProcessor
 
 
 class AbstractArrayJoinOptimizer(ClickhouseQueryProcessor):
