+++ snuba/query/processors/physical/mandatory_condition_applier.py
@@ -1,6 +1,6 @@
-from snuba.clickhouse.processors import ClickhouseQueryProcessor
 from snuba.clickhouse.query import Query
 from snuba.query.conditions import combine_and_conditions
+from snuba.query.processors.physical.processors import ClickhouseQueryProcessor
 from snuba.query.query_settings import QuerySettings
