+++ snuba/query/processors/physical/null_column_caster.py
@@ -1,11 +1,11 @@
 from typing import Dict, Sequence
 
 from snuba.clickhouse.columns import FlattenedColumn, SchemaModifiers
-from snuba.clickhouse.processors import ClickhouseQueryProcessor
 from snuba.clickhouse.query import Query
 from snuba.datasets.storage import ReadableTableStorage
 from snuba.query.expressions import Column, Expression, FunctionCall, Literal
 from snuba.query.functions import AGGREGATION_FUNCTIONS
+from snuba.query.processors.physical.processors import ClickhouseQueryProcessor
 from snuba.query.query_settings import QuerySettings
