+++ snuba/query/processors/physical/prewhere.py
@@ -1,7 +1,6 @@
 from typing import Optional, Sequence, Set
 
 from snuba import environment, settings
-from snuba.clickhouse.processors import ClickhouseQueryProcessor
 from snuba.clickhouse.query import Query
 from snuba.query.accessors import get_columns_in_expression
 from snuba.query.conditions import (
@@ -10,6 +9,7 @@ from snuba.query.conditions import (
     get_first_level_and_conditions,
 )
 from snuba.query.expressions import FunctionCall
+from snuba.query.processors.physical.processors import ClickhouseQueryProcessor
 from snuba.query.query_settings import QuerySettings
 from snuba.utils.metrics.wrapper import MetricsWrapper
