+++ snuba/query/processors/physical/table_rate_limit.py
@@ -1,7 +1,7 @@
 from typing import Optional
 
-from snuba.clickhouse.processors import ClickhouseQueryProcessor
 from snuba.clickhouse.query import Query
+from snuba.query.processors.physical.processors import ClickhouseQueryProcessor
 from snuba.query.query_settings import QuerySettings
 from snuba.state import get_configs
 from snuba.state.rate_limit import TABLE_RATE_LIMIT_NAME, RateLimitParameters
