+++ tests/query/processors/test_table_rate_limit.py
@@ -1,9 +1,9 @@
 import pytest
 
 from snuba.clickhouse.columns import ColumnSet
-from snuba.clickhouse.processors import ClickhouseQueryProcessor
 from snuba.clickhouse.query import Query
 from snuba.query.data_source.simple import Table
+from snuba.query.processors.physical.processors import ClickhouseQueryProcessor
 from snuba.query.processors.physical.table_rate_limit import TableRateLimit
 from snuba.query.query_settings import HTTPQuerySettings
 from snuba.state import set_config
